Working Prototype with style, check tournament rules

+++ calculator/models.py
@@ -3,10 +3,10 @@ from django.db import models
 
 class Product(models.Model):
     name = models.CharField(max_length = 30)
-    calories = models.IntegerField()
-    bilky = models.IntegerField()
-    zhury = models.IntegerField()
-    vuglevody = models.IntegerField()
+    calories = models.FloatField()
+    bilky = models.FloatField()
+    zhury = models.FloatField()
+    vuglevody = models.FloatField()
 
     def __str__(self):
         return self.name
